refactor: handler 상수 분리 및 핸들러 분리
