+++ src/api/mocks/handlers/index.ts
@@ -1,0 +1,15 @@
+import { getProductsHandler } from './productHandlers';
+import {
+  getCartHandler,
+  postCartHandler,
+  deleteCartHandler,
+  patchCartHandler,
+} from './cartHandlers';
+
+export const handlers = [
+  getProductsHandler,
+  getCartHandler,
+  postCartHandler,
+  deleteCartHandler,
+  patchCartHandler,
+];
